+++ src/routes/Tomogram.tsx
@@ -15,85 +15,51 @@ import {
   Skeleton,
 } from "@chakra-ui/react";
 import { useCallback, useEffect, useState } from "react";
-import { useNavigate, useParams, useSearchParams } from "react-router-dom";
-import { client } from "../utils/api/client";
+import { useLoaderData, useNavigate, useParams, useSearchParams } from "react-router-dom";
 import { Tomogram } from "../components/tomogram/main";
-import { parseData } from "../utils/generic";
 import { CollectionData } from "../schema/interfaces";
 import { MotionPagination } from "../components/motion/pagination";
 import { InfoGroup } from "../components/visualisation/infogroup";
-import { buildEndpoint } from "../utils/api/endpoint";
-import { collectionConfig } from "../utils/config/parse";
 import { MdList, MdRedo } from "react-icons/md";
 import { components } from "../schema/main";
 
-type ProcessingResponse = components["schemas"]["ProcessingJobOut"];
+type ProcessingJob = components["schemas"]["ProcessingJobOut"];
 
 const TomogramPage = () => {
   const params = useParams();
+  const loaderData = useLoaderData() as {
+    collection: CollectionData;
+    total: number;
+    page: number;
+    jobs: ProcessingJob[] | null;
+  };
   const [searchParams, setSearchParams] = useSearchParams();
-  const [jobs, setJobs] = useState<ProcessingResponse[] | null | undefined>();
-  const [collectionData, setCollectionData] = useState<CollectionData>({ info: [], comments: "" });
-  const [pageCount, setPageCount] = useState(1);
-  const [onlyProcessed, setOnlyProcessed] = useState(searchParams.get("onlyProcessed") === "true");
+  const [onlyProcessed, setOnlyProcessed] = useState(searchParams.get("onlyTomograms") === "true");
   const [accordionIndex, setAccordionIndex] = useState<number | number[]>(0);
   const navigate = useNavigate();
 
   const updateCollection = useCallback(
     (page: number) => {
-      navigate(`../${page}`, { relative: "path" });
+      navigate(`../${page}?onlyTomograms=${onlyProcessed}`, { relative: "path" });
     },
-    [navigate]
+    [navigate, onlyProcessed]
   );
 
   const updateTomogramFilter = useCallback(() => {
     setOnlyProcessed(!onlyProcessed);
-    setSearchParams({ onlyProcessed: (!onlyProcessed).toString() });
+    setSearchParams({ onlyTomograms: (!onlyProcessed).toString() });
   }, [onlyProcessed, setSearchParams]);
 
   useEffect(() => {
     document.title = `PATo » Tomograms » ${params.collectionIndex}`;
-
-    /** There should be 3 possible states: a null tomogram (for when it is still being processed),
-    /* and undefined tomogram (waiting for information client-side) and a valid tomogram */
-    setJobs(undefined);
-    client
-      .safe_get(
-        `${buildEndpoint(
-          "dataCollections",
-          params,
-          1,
-          parseInt(params.collectionIndex ?? "1")
-        )}&onlyTomograms=${onlyProcessed}`
-      )
-      .then((response) => {
-        if (response.data.total && response.data.items) {
-          setPageCount(response.data.total);
-          if (params.collectionIndex && params.collectionIndex > response.data.total) {
-            updateCollection(1);
-          }
-          setCollectionData(parseData(response.data.items[0], collectionConfig) as CollectionData);
-
-          client
-            .safe_get(`dataCollections/${response.data.items[0].dataCollectionId}/processingJobs?limit=3`)
-            .then((response) => {
-              if (response.status === 200 && response.data) {
-                const items = response.data.items as ProcessingResponse[];
-                setJobs(items);
-              } else {
-                setJobs(null);
-              }
-            });
-        }
-      });
-  }, [params, navigate, updateCollection, onlyProcessed]);
+  }, [params]);
 
   return (
     <Box>
       <HStack marginBottom={2}>
         <VStack w='100%'>
           <HStack w='100%'>
-            <Heading>{collectionData.comments}</Heading>
+            <Heading>{loaderData.collection.comments}</Heading>
             <Tag colorScheme='teal'>Tomogram</Tag>
             <Spacer />
             <Tooltip label='Run Reprocessing'>
@@ -111,7 +77,7 @@ const TomogramPage = () => {
               size='md'
               onChange={updateCollection}
               displayDefault={parseInt(params.collectionIndex ?? "1")}
-              total={pageCount}
+              total={loaderData.total}
             />
           </HStack>
           <HStack w='100%'>
@@ -126,11 +92,11 @@ const TomogramPage = () => {
           </HStack>
         </VStack>
       </HStack>
-      <InfoGroup cols={3} info={collectionData.info} />
+      <InfoGroup cols={3} info={loaderData.collection.info} />
       <Divider my={2} />
-      {jobs ? (
+      {loaderData.jobs ? (
         <Accordion onChange={setAccordionIndex} index={accordionIndex} allowToggle>
-          {jobs.map((job, i) => (
+          {loaderData.jobs.map((job, i) => (
             <Tomogram
               key={job.AutoProcProgram.autoProcProgramId}
               autoProc={job.AutoProcProgram}